requirements.txt
torch>=1.7.0
torchvision>=0.8.0
timm>=0.4.12
transformers>=4.0.0
scikit-learn>=0.24.0
tensorly>=0.6.0
numpy>=1.18.0
PyYAML>=5.3.1
wandb>=0.12.0

